+++ aa_forum/constants.py
@@ -14,8 +14,8 @@ verbose_name_slug: str = slugify(verbose_name, allow_unicode=True)
 user_agent: str = f"{verbose_name_slug} v{__version__} {github_url}"
 
 # Setting keys
-SETTING_MESSAGESPERPAGE = "default_max_messages"
-SETTING_TOPICSPERPAGE = "default_max_topics"
+SETTING_MESSAGESPERPAGE = "messages_per_page"
+SETTING_TOPICSPERPAGE = "topics_per_page"
 
 # All internal URLs need to start with this prefix
 # to prevent conflicts with user generated forum URLs